Move `ToGridFSUploadPublisherUnit` such that it's not global; add the test proposed by Ross
JAVA-4303
